categoryCard fixed
