+++ src/components/categoryCard/CategoryCard.jsx
@@ -1,27 +1,46 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import styles from './CategoryCard.module.scss';
-import { useTheme } from '../../context/ThemeContext';
+import { useSelector, useDispatch } from 'react-redux';
+import { fetchCategories } from '../../utils/fetchClient';
 import { BASE_URL } from '../../constants/constants';
+import styles from './CategoryCard.module.scss';
+import Container from '../container/Container';
 
-export const CategoryCard = ({ title, image, id, limit }) => {
-  const limitClass = limit === 4 ? styles['limit-4'] : styles['limit-5'];
-  const { isDarkTheme } = useTheme();
+const CategoryCard = ({ limit }) => {
+  const dispatch = useDispatch();
+  const { categories, loading, error } = useSelector(
+    (state) => state.categories
+  );
+
+  useEffect(() => {
+    dispatch(fetchCategories());
+  }, [dispatch]);
+
+  if (loading) return <div>Loading...</div>;
+  if (error) return <div>{error}</div>;
+  if (!categories || categories.length === 0)
+    return <div>No categories available.</div>;
+
   return (
-    <div className={`${styles.categoriesItemWrapper} ${limitClass}`}>
-      <div className={styles.imageContainer}>
-        <img
-          src={`${BASE_URL}${image}`}
-          alt={`Category ${title}`}
-          className={styles.categoriesImage}
-        />
+      <div className={styles.categoriesWrapper}>
+        {categories.slice(0, limit).map((item) => (
+          <div key={item.id} className={styles.categoriesItemWrapper}>
+            <div className={styles.imageContainer}>
+              <img
+                src={`${BASE_URL}${item.image}`}
+                alt={`Category ${item.title}`}
+                className={styles.categoriesImage}
+              />
+            </div>
+            <div className={styles.categoriesTitle}>
+              <Link className={styles.link} to={`/categories/${item.id}`}>
+                <span>{item.title}</span>
+              </Link>
+            </div>
+          </div>
+        ))}
       </div>
-
-      <div className={styles.categoriesTitle}>
-        <Link className={styles.link} to={`/categories/${id}`}>
-          <h3 className={styles.categoriesTitle}>{title}</h3>
-        </Link>
-      </div>
-    </div>
   );
 };
+
+export default CategoryCard;
