+++ src/components/categoryCard/CategoryCard.module.scss
@@ -1,46 +1,52 @@
-.categoriesItemWrapper {
+@use '../../assets/styles/mixin' as *;
+
+.categoriesWrapper {
+  padding-top: 20px;
   display: flex;
-  flex-direction: column;
-  flex: 1 1 320px;
   flex-wrap: wrap;
-  gap: 16px;
-  padding-bottom: 20px;
+  gap: 24px;
+  padding-bottom: 25px;
   max-width: 100%;
-  text-align: center;
-  cursor: pointer;
-  transition: transform 0.4s ease-in-out;
-
-  &:hover {
-    transform: scale(1.01);
-    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.2);
+  @include on-laptop {
+    flex-wrap: nowrap;
   }
 
-  &.limit-4:nth-child(n + 5) {
-    display: none;
-  }
-
-  &.limit-5:nth-child(n + 6) {
-    display: none;
-  }
-
-  .categoriesTitle {
+  .categoriesItemWrapper {
+    display: flex;
+    flex-direction: column;
+    flex: 1 1 300px;
+    flex-wrap: wrap;
+    gap: 16px;
+    padding-bottom: 20px;
     text-align: center;
-    font-size: 20px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 130%;
-  }
+    max-width: 400px;
+    cursor: pointer;
+    transition: transform 0.4s ease-in-out;
 
-  .imageContainer {
-    width: 100%;
-    max-width: 400px; /* Максимальная ширина картинки */
-  }
+    &:hover {
+      transform: scale(1.01);
+      box-shadow: 0 6px 12px rgba(0, 0, 0, 0.2);
+    }
 
-  .categoriesImage {
-    width: 100%;
-    height: 390px; /* Высота картинки */
-    object-fit: cover;
-    border-radius: 5px;
-    display: block;
+    .categoriesTitle {
+      text-align: center;
+      font-size: 20px;
+      font-style: normal;
+      font-weight: 500;
+      line-height: 130%;
+    }
+
+    .imageContainer {
+      max-width: 100%;
+      height: 390px; /* Высота картинки */
+    }
+
+    .categoriesImage {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+      border-radius: 5px;
+      display: block;
+    }
   }
 }
