--- src/pages/categories/Categories.module.scss
@@ -1,12 +1,0 @@
-@use '../../assets/styles/mixin' as *;
-
-.categoriesWrapper {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 24px;
-  padding-bottom: 25px;
-
-  @include on-laptop {
-    flex-wrap: nowrap;
-  }
-}
