+++ src/pages/home/Home.jsx
@@ -6,7 +6,7 @@ import Discount from '../../components/Discount/Discount.jsx';
 import MainSectionTitle from '../../components/mainSectionTitle/MainSectionTitle.jsx';
 import CategoryProducts from '../categoryProducts/CategoryProducts.jsx';
 import Categories from '../categories/Categories.jsx';
-import { CategoryCard } from '../../components/categoryCard/CategoryCard.jsx';
+import CategoryCard from '../../components/categoryCard/CategoryCard.jsx';
 
 export default function Home() {
   return (
@@ -18,8 +18,9 @@ export default function Home() {
           to="/categories "
           text="All categories"
         />
+        <CategoryCard limit={4} />
       </Container>
-      <Categories limit={4} />
+
       <Discount />
       <SaleSection />
     </>
